Refactoring to be considered before adding MMTk
